use raw pointers instead of ref, in and out parameters.
this erases the possibility of writing incorrect bindings (a misinterpretation of X parameter in Y function, for example).

the use of the bindings is pretty much the same (except that basically everything has to be unsafe, but that's not an actual problem, the wrapper itself is unsafe lol).

this is easier to mantain, because it removes the necessity of having like 300 overloads.

custom marshalling for strings and booleans is kept.

-removed every instance of the in, ref and out keyword in functions in favor of using raw pointers instead.
-added ConstAttribute to mark a parameter is passed as read-only in C.
-array parameters are now pointers, letting you have more control over the unmanaged memory.
this also applies to fields.
-added UnmanagedToManagedString/s() to make it easier to marshall unmanaged strings and array of strings.
-updated documentation according to this change, also